fix joboperations and tests
